Ignore Struct Keys From Being Stringified
Because Elixir structs can represent complex data types via Protocols
their key/value pairs are not enumerable.  This leads to problems when
attempting to stringify maps that may hold some structs in it.  Before
this commit the behavior of the code was it would raise an exception
when it met such data:

```
** (Protocol.UndefinedError) protocol Enumerable not implemented for #Decimal<7>
```

Instead of failing this elects to leave structs alone as they do not
function the same as normal maps but unforunately pass such guard
clauses as `is_map`.